refactor result context
